Updated 'Centralised_system.html'.
The way in which margins and paddings are implemented has been improved. Additionally, the 'Employee Handbook' line is now on one line (as opposed to two as in the past).

Note: the previous commit means that commits prior to the one before it are no longer functional, because the file that it used to generate both the Wiki and Employee Handbook logos was changed.