Updated 'Centralised_system.html'.
Improved the visual appearance of the format of the application. Its spacing/padding/margins are now all equal size. Its vertical scroll bar begins at the top of the page, and does not allow the user to scroll past a specified point at the beginning or ending of the page on which the application resides.