Updated 'Centralised_system.html'.
Has made the files 'Android_Chrome.css', 'Samsung_Internet.css', 'iPhone_Chrome.css', 'iPhone_Safari.css' obsolete by implementing one style sheet. However, there is variation with in the 'Xero Me' and 'vGo' hyperlinks links to according to the operating system of the device on which the application is running on.