Working version of 'index.html'.
The file named 'index.html' is the primary application interface, that provides a grid of hyperlinks to several application hyperlinks. That grid layout is responsive. It provides deep linking support for native apps (for both IO and Android devices). It directs its users to the app store pages of the respective applications of those which are uninstalled (for both IO and Android devices). It provides Progressive Web App (PWA) support for installation onto its users' devices' home screens. It is optimised for mobile devices.