Working version of 'manifest.json'.
The file named 'manifest.json' enables Progressive Web App (PWA) functionality. It defines the names of its applications' hyperlinks, their icons, as well as their themes. It facilities a "Add to Home Screen" functionality. It determines the display mode and orientation preferences of the application. It configures its applications' hyperlinks.